refactor(rag-knowledge): U2/U3 残課題完遂 — RAGKnowledgeService 削除 + テストファイル Port 単位分割 (#730)
Issue #730 の Q3=B「ファサード削除」と Q7=B「テストファイルも Port 単位に分割」
の残課題を完遂する。

`RAGKnowledgeService` 削除（Q3=B）:

- `src/rag/rag_knowledge.py` を削除
  - thin facade (約 230 LOC) は新 Port + Adapter 群への委譲のみだったため、
    呼び出し元更新が完了した時点で削除可能だった
- `tests/factories.py`:
  - `make_rag_knowledge_service` を削除
  - `make_search_adapter` / `make_source_management_adapter` /
    `make_stats_adapter` の 3 ファクトリを新設

テストファイル Port 単位分割（Q7=B）:

Service 系（`tests/test_rag_knowledge.py` 802 LOC + `test_rag_knowledge_hybrid.py`
を分割）:
- `tests/test_search_port.py` — TestRetrieve / TestRAGDebugLog /
  TestRAGRetrievalResultSources / TestSimilarityThreshold /
  TestRetrieveRawResults / TestConfiguration
- `tests/test_search_port_hybrid.py` — TestHybridSearchDisabled /
  TestHybridSearchEnabled / TestTableDataSearch / TestKeywordExactMatch /
  TestHybridSearchEngineInitialization
- `tests/test_source_management_port.py` — TestDeleteSource /
  TestGetFullPageText（BM25 連動 delete テスト統合）
- `tests/test_stats_port.py` — TestGetStats
- `tests/test_smart_chunking.py` — `smart_chunk` 関数の単体テスト
  （Service の `_smart_chunk` proxy ではなく関数直接テストに変更）

Pipeline 系（`tests/test_pipeline_controller.py` 1908 LOC を分割）:
- `tests/test_change_detector.py` — TestClassifyChanges /
  TestClassifyChangesMedia / TestScanAllAsAdded（約 350 LOC 抽出）
- `tests/test_pipeline_controller.py` 1908 → 1672 LOC（orchestration テストに集中）

旧テストファイルの削除:
- `tests/test_rag_knowledge.py` 削除
- `tests/test_rag_knowledge_hybrid.py` 削除

各 Adapter 単体テストはフィクスチャを `make_*_adapter()` ファクトリで
構築し、Adapter のメソッドを直接呼ぶ。Service 経由の `_smart_chunk` /
`_ingest_crawled_page` 関連テストは smart_chunk 関数テスト + cli の
`_ingest_page_for_testing` テストに置換済み（U2 の commit `08666e9`）。

Pipeline 系は ChangeDetector のみ抽出。ChangeHandler / MetadataBuilder の
単体テストは元々存在せず、Controller orchestration テスト経由でカバーされて
いるため抽出対象なし（必要に応じて follow-up Issue で対応）。

検証:
- ruff / mypy 0 error（src/rag 131 source files）
- pytest 全 2024 件 pass
- markdownlint 0 error
