+++ tests/test_pipeline_controller.py
@@ -1341,242 +1341,6 @@ class TestPipelineHistory:
         assert len(history) == 2
 
 
-class TestClassifyChanges:
-    """_classify_changes のテスト."""
-
-    def test_data_and_meta_together(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-    ) -> None:
-        ctrl, _, _ = controller
-        raw = [
-            ("M", "web/example.com/page.html", ""),
-            ("M", "web/example.com/page.html.meta", ""),
-        ]
-        changes = ctrl._classify_changes(raw)
-        assert len(changes) == 1
-        assert changes[0].file_path == "web/example.com/page.html"
-        assert changes[0].status.value == "modified"
-
-    def test_meta_only(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-    ) -> None:
-        ctrl, _, _ = controller
-        raw = [
-            ("M", "web/example.com/page.html.meta", ""),
-        ]
-        changes = ctrl._classify_changes(raw)
-        assert len(changes) == 1
-        assert changes[0].status.value == "meta_only"
-        assert changes[0].file_path == "web/example.com/page.html"
-
-    def test_gitignore_filtered(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-    ) -> None:
-        ctrl, _, _ = controller
-        raw = [
-            ("M", ".gitignore", ""),
-            ("A", "local/test.txt", ""),
-        ]
-        changes = ctrl._classify_changes(raw)
-        assert len(changes) == 1
-        assert changes[0].file_path == "local/test.txt"
-
-
-class TestClassifyChangesMedia:
-    """_classify_changes の media 親 JSON 検出テスト."""
-
-    def test_media_added_triggers_parent_json_and_excludes_self(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-        workspace: dict[str, Path],
-    ) -> None:
-        """media 追加時、親 JSON のみが ChangeEntry として返ることを検証する（#597）."""
-        ctrl, _, _ = controller
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/rkey1.json",
-            '{"post": {}}',
-        )
-        raw = [
-            ("A", "bluesky/did/2026/04/media/rkey1/image_0.webp", ""),
-        ]
-        changes = ctrl._classify_changes(raw)
-        paths = {c.file_path: c.status.value for c in changes}
-        assert "bluesky/did/2026/04/media/rkey1/image_0.webp" not in paths
-        assert paths == {"bluesky/did/2026/04/rkey1.json": "modified"}
-
-    def test_media_added_no_parent_json_file_yields_no_changes(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-    ) -> None:
-        """親 JSON が存在しない孤児 media は ChangeEntry を生成しない（#597）."""
-        ctrl, _, _ = controller
-        raw = [
-            ("A", "bluesky/did/2026/04/media/rkey1/image_0.webp", ""),
-        ]
-        changes = ctrl._classify_changes(raw)
-        paths = {c.file_path for c in changes}
-        assert "bluesky/did/2026/04/media/rkey1/image_0.webp" not in paths
-        assert "bluesky/did/2026/04/rkey1.json" not in paths
-
-    def test_media_modified_with_existing_parent_triggers_parent_only(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-        workspace: dict[str, Path],
-    ) -> None:
-        """media 変更時、親 JSON のみ MODIFIED として返り media 自身は除外される（#597）."""
-        ctrl, _, _ = controller
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/rkey1.json",
-            '{"post": {}}',
-        )
-        raw = [
-            ("M", "bluesky/did/2026/04/media/rkey1/image_0.webp", ""),
-        ]
-        changes = ctrl._classify_changes(raw)
-        paths = {c.file_path: c.status.value for c in changes}
-        assert "bluesky/did/2026/04/media/rkey1/image_0.webp" not in paths
-        assert paths == {"bluesky/did/2026/04/rkey1.json": "modified"}
-
-    def test_parent_json_already_in_diff(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-        workspace: dict[str, Path],
-    ) -> None:
-        ctrl, _, _ = controller
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/rkey1.json",
-            '{"post": {}}',
-        )
-        raw = [
-            ("A", "bluesky/did/2026/04/media/rkey1/image_0.webp", ""),
-            ("M", "bluesky/did/2026/04/rkey1.json", ""),
-        ]
-        changes = ctrl._classify_changes(raw)
-        # 親 JSON が既に diff にある場合、重複追加されない
-        json_entries = [c for c in changes if c.file_path == "bluesky/did/2026/04/rkey1.json"]
-        assert len(json_entries) == 1
-        assert json_entries[0].status.value == "modified"
-
-    def test_media_renamed_across_rkey_triggers_both_parents(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-        workspace: dict[str, Path],
-    ) -> None:
-        """media が別 rkey へリネームされた場合、新旧両方の親 JSON が再変換対象になる（#597 Copilot 指摘）."""
-        ctrl, _, _ = controller
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/rkey_old.json",
-            '{"post": {}}',
-        )
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/rkey_new.json",
-            '{"post": {}}',
-        )
-        raw = [
-            (
-                "R",
-                "bluesky/did/2026/04/media/rkey_new/image_0.webp",
-                "bluesky/did/2026/04/media/rkey_old/image_0.webp",
-            ),
-        ]
-        changes = ctrl._classify_changes(raw)
-        paths = {c.file_path: c.status.value for c in changes}
-        assert "bluesky/did/2026/04/media/rkey_new/image_0.webp" not in paths
-        assert "bluesky/did/2026/04/media/rkey_old/image_0.webp" not in paths
-        assert paths.get("bluesky/did/2026/04/rkey_old.json") == "modified"
-        assert paths.get("bluesky/did/2026/04/rkey_new.json") == "modified"
-
-    def test_media_renamed_to_nonmedia_preserves_new_path_processing(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-        workspace: dict[str, Path],
-    ) -> None:
-        """media → 非 media へのリネーム時、旧親 JSON は再変換対象、新パスは通常通り登録される."""
-        ctrl, _, _ = controller
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/rkey_old.json",
-            '{"post": {}}',
-        )
-        raw = [
-            (
-                "R",
-                "local/moved_image.webp",
-                "bluesky/did/2026/04/media/rkey_old/image_0.webp",
-            ),
-        ]
-        changes = ctrl._classify_changes(raw)
-        paths = {c.file_path: c.status.value for c in changes}
-        assert paths.get("bluesky/did/2026/04/rkey_old.json") == "modified"
-        assert paths.get("local/moved_image.webp") == "renamed"
-
-
-class TestScanAllAsAdded:
-    """_scan_all_as_added のテスト."""
-
-    def test_excludes_bluesky_media(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-        workspace: dict[str, Path],
-    ) -> None:
-        """初回スキャン経路で bluesky media が独立 ChangeEntry として返らない（#597）."""
-        ctrl, _, _ = controller
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/rkey1.json",
-            '{"post": {}}',
-        )
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/media/rkey1/image_0.webp",
-            "dummy",
-        )
-        _place_local_file(
-            workspace["source"],
-            "bluesky/did/2026/04/media/rkey1/video_0.ts",
-            "dummy",
-        )
-        ctrl.init_repo()
-        ctrl.commit("initial")
-
-        entries = ctrl._scan_all_as_added()
-        paths = {e.file_path for e in entries}
-        assert "bluesky/did/2026/04/rkey1.json" in paths
-        assert "bluesky/did/2026/04/media/rkey1/image_0.webp" not in paths
-        assert "bluesky/did/2026/04/media/rkey1/video_0.ts" not in paths
-
-    def test_excludes_meta_and_sidecar_files(
-        self,
-        controller: tuple[PipelineController, StubConverter, StubIndexer],
-        workspace: dict[str, Path],
-    ) -> None:
-        """初回スキャンで .meta・sidecar（aozora/catalog.csv 等）も除外する.
-
-        除外ロジックは `is_source_file` に一元化されており、ここでは
-        独立ソース以外が ChangeEntry に混入しないことを検証する。
-        """
-        ctrl, _, _ = controller
-        _place_local_file(workspace["source"], "local/a.txt", "content")
-        _place_local_file(workspace["source"], "local/a.txt.meta", "{}")
-        _place_local_file(workspace["source"], "aozora/catalog.csv", "data")
-        ctrl.init_repo()
-        ctrl.commit("initial")
-
-        entries = ctrl._scan_all_as_added()
-        paths = {e.file_path for e in entries}
-        assert "local/a.txt" in paths
-        assert "local/a.txt.meta" not in paths
-        assert "aozora/catalog.csv" not in paths
-
-
 class TestRenamedLocalSourceId:
     """local リネーム時の source_id 更新テスト."""
